src/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord.py
idna==2.9
multidict==4.7.6
numpy==1.19.0
pandas==1.0.5
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
python-dotenv==0.13.0